requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-lru==1.0.2
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
greenlet==1.1.1
idna==3.2
multidict==5.1.0
psycopg2-binary==2.9.1
pyaes==1.6.1
Pyrogram==2.0.30
PySocks==1.7.1
python-decouple==3.6
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
TgCrypto==1.2.3
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
yarl==1.6.3
